refactor: split JSON schema handler in its own class
