+++ src/Event/BootstrapEventHandler.php
@@ -1,22 +1,29 @@
 <?php
+/**
+ * BEdita, API-first content management framework
+ * Copyright 2022 Atlas Srl, Chialab Srl
+ *
+ * This file is part of BEdita: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * See LICENSE.LGPL or <http://gnu.org/licenses/lgpl-3.0.html> for more details.
+ */
 
 namespace BEdita\Placeholders\Event;
 
 use BEdita\API\Controller\ObjectsController;
-use BEdita\Core\Model\Entity\ObjectType;
 use BEdita\Core\Model\Table\ObjectsBaseTable;
 use BEdita\Core\Model\Table\ObjectsTable;
 use Cake\Event\Event;
 use Cake\Event\EventListenerInterface;
-use Cake\ORM\Locator\LocatorAwareTrait;
 
 /**
  * Attach placeholders behavior and component to relevant models and controllers, respectively, upon initialization.
  */
 class BootstrapEventHandler implements EventListenerInterface
 {
-    use LocatorAwareTrait;
-
     /**
      * @inheritDoc
      */
@@ -25,7 +32,6 @@ class BootstrapEventHandler implements EventListenerInterface
         return [
             'Model.initialize' => 'onModelInitialize',
             'Controller.initialize' => 'onControllerInitialize',
-            'ObjectType.getSchema' => 'onGetSchema',
         ];
     }
 
@@ -59,31 +65,5 @@ class BootstrapEventHandler implements EventListenerInterface
         }
 
         $controller->loadComponent('BEdita/Placeholders.Placeholders');
-    }
-
-    /**
-     * Modify object type schema by marking fields where placeholders are read from.
-     *
-     * @param \Cake\Event\Event $event Dispatched event.
-     * @param array $schema Automatically generated JSON schema.
-     * @param \BEdita\Core\Model\Entity\ObjectType $objectType Object type.
-     * @return array
-     */
-    public function onGetSchema(Event $event, array $schema, ObjectType $objectType): array
-    {
-        $table = $this->getTableLocator()->get($objectType->table);
-        if (!$table->hasBehavior('Placeholders')) {
-            return $schema;
-        }
-
-        $fields = $table->getBehavior('Placeholders')->getConfig('fields');
-        foreach ($fields as $field) {
-            if (!isset($schema['properties'][$field])) {
-                continue;
-            }
-            $schema['properties'][$field]['placeholders'] = true;
-        }
-
-        return $schema;
     }
 }
