+++ src/Plugin.php
@@ -14,6 +14,7 @@
 namespace BEdita\Placeholders;
 
 use BEdita\Placeholders\Event\BootstrapEventHandler;
+use BEdita\Placeholders\Event\JsonSchemaEventHandler;
 use Cake\Core\BasePlugin;
 use Cake\Core\PluginApplicationInterface;
 use Cake\Event\EventManager;
@@ -31,6 +32,7 @@ class Plugin extends BasePlugin
         parent::bootstrap($app);
 
         EventManager::instance()
-            ->on(new BootstrapEventHandler());
+            ->on(new BootstrapEventHandler())
+            ->on(new JsonSchemaEventHandler());
     }
 }
